[Rust SDK] Add ProgressStorage trait

+++ rust/processor/src/processors/coin_processor.rs
@@ -264,7 +264,7 @@ impl ProcessorTrait for CoinProcessor {
         transactions: Vec<Transaction>,
         start_version: u64,
         end_version: u64,
-        _: Option<u64>,
+        _: Option<u8>,
     ) -> anyhow::Result<ProcessingResult> {
         let processing_start = std::time::Instant::now();
         let mut conn = self.get_conn().await;
