+++ rust/processor/src/processors/events_processor.rs
@@ -116,7 +116,7 @@ impl ProcessorTrait for EventsProcessor {
         transactions: Vec<Transaction>,
         start_version: u64,
         end_version: u64,
-        _: Option<u64>,
+        _: Option<u8>,
     ) -> anyhow::Result<ProcessingResult> {
         let processing_start = std::time::Instant::now();
         let mut conn = self.get_conn().await;
